Proposal for expanding the ADR committee
Conversation in the gaia discord seems to merit an update to the ADR committee. I haven't spoken to any of the individuals to confirm their interest or availability but I'd like to use this as a more productive forum to push the conversation further. Especially looking for comment from those proposed (@marbar3778 @ebuchman @robert-zaremba @amaurym @zmanian @sunnya97 @fedekunze @odeke-em  )